Fix object garbage collection processing
Now, GC processing is done in free.  That is not appropriate, because

a) Some input argument need to be freed, but not GC-ed
b) GC need to access values instead of memory location

So added individual GC function in converter/translator/arg-processor
and call them after function call.

For object GC processing, when caller get an object from return value
or out argument of a function, if the object ownership is transferred,
the reference count should have been increased for caller already.  So
to setup GC for an object, it should be enough only to call
g-object-ref-sink if it is floating to get the ownership.