Garbage Collection for "transfer none"
For "transfer none" return value and out parameters, the reference
count of returned object will not be increased for the caller, but a
new pointer is established in lisp for it.  For example,
gtk_bin_get_child.  For these objects, we should increase reference
count when the new pointer is established in lisp and decrease
reference count when the pointer is garbage collected.