Garbage collection of info-ffi for "transfer none"
For "transfer none" returned info-ffi, the reference count of returned
info-ffi will not be increased for the caller, but a new pointer is
established in lisp for it.  For example, g-base-info-get-container.
For these info-ffis, we should increase reference count when the new
pointer is established in lisp and decrease reference count when the
pointer is garbage collected.