Represent foreign type with CLOS
Originally tried to do this via extending CFFI foreign types, but some
features are missing in CFFI, such as freeing aggregated element of
array.

One of the best part of the patch is to represent '(union argument) as
a foreign type, so that we can just use normal foreign type function
on that instead of define another family of functions.