src/function.lisp: fix build-function to handle structs as "inout" args without crashing lisp
the `with-foreign-alloc'ed `values-out' in build-function was being
used incorrectly with inout arguments. fix a memory-overflow in
in-arg-setup by always allocating memory, before letting in-arg-setup
stomp it.

* src/function.lisp: (mem-alloc): methods for struct-type and
union-type. (make-arg): allocate memory at voutp on values-out array,
and point the giargs-in and giargs-out to it.

* src/function.lisp: (out-arg->value) try to free the giarg allocated
for an "inout" argument before returning.

NOTE: out-args are not generally freed
except for what happens via mem-get in not out-arg->value.
This needs further work.

Also see the commit message in
c27fdb6740389250e16e9e9bf93f8f285b04dc74